feat: set metadata `validUntil` to X.509 certificate expiry datetime
Currently, `validUntil` is set to two days by default, thanks to the
underlying PHP SAML library.  This means that the `validUntil`
attribute must be removed or changed prior to handoff to the IdP.

Instead, parse the X.509 certificate provided in settings, and use the
parsed expiry date as the value for `validUntil`.  Considering the default
10-year lifespan of certificates generated with `openssl x509`, this
approach not only gives the metadata plenty of time to live, but it
also makes the `validUntil` attribute more accurate.